+++ precision.h
@@ -1,7 +1,7 @@
 #ifndef PRECISION_H
 #define PRECISION_H
 
-typedef int index_t;
+typedef long long index_t;
 typedef float coord_t;
 
 #endif /* PRECISION_H */
